Added rich text editing for news stories
Adding more TODOs.
Preserving changes to news stories by not reloading them from the
hardcoded values, (hacked in NewsAPI for now.)
Fixed a warning/formatting issue on Fixtures page.
Fixed Nav links to change the root url.
Added title to index page.
Added some notes to the readme that I need to ask about.
Added compression to the express server, (not sure if it is working)
based on
https://github.com/reactjs/react-router-tutorial/tree/master/lessons/11-
productionish-server