XML to JSON Code
Taking some of the news xml files and writing code to convert it to
JSON. This is all of them except for a) conflicting filenames and b)
those in the .zip file within the Batch1/News folder. Need to add those
to filenames.js.
Need to upload it to the server to store in Firebase.