Getting setup for editing single pages, such as Tickets
Pages will represent single page content, such as the Tickets page,
which will still need to be edited.
Adding Actions, API, Model and Route.
Cleaned up the way reducers are exported/imported and therefore how the
store is configured.

Still need to add the Pages Edit components, add those to the App/React
Router, put links on the Admin form and then update the content of the
individual page components, such as Tickets.