Added ability to add and edit players
Also fixed a bug to allow creation of new Pages.
Tweaked rendering of PlayerDetail so the biography renders as formatted
text, not raw HTML.
Tweaked PlayersAPI so players are also sorted by last name, mostly for
the Teams component, but also the PlayersEdit.
Only show Active players on the Teams page.
Now consistently storing DOB as a number in MS, not a date string.