Adding collapsed Nav for medium screens
Too many links on medium screens now, so collapsed that down.
Removing a sponsor from Home screen.
Adding another tower ad on fixtures screen.
Changing link on Nav to Club Shop to be a <a> instead of a <Link>.
Cleaning up the background color on ads, since this isn’t needed now
that they are being shown.