Replacing Lazy Loading code for images
react-lazy-image wasn't working in ie11/Edge, so I replaced it with the
more generic react-lazyload.

Updated and tested all relevant components.

Remaining issue is with the Team page, which will need to be reworked
since the images will not load without scrolling the page, (not the
container div.) Could look at using forceCheck() when scrolling the
container: https://github.com/jasonslyvia/react-lazyload#forcecheck