Almost done with Sponsors
Need to make some tweaks to get the admin functions fully working, (such as Delete,) since the FirebaseCacheSponsors doesn't remove items. Probably need to restructure the firebase data for sponsors to have another level under it, (e.g. sponsors:sponsros:<id>:<params>,) so that the cache gets the full list of sponsors in the snapshot. Then test and add full data to DB.