Removing recent cache
While the generally cache works and saves effort on the server, there is an edge case where future news posts are not taken into account once their embargo time expires. They wouldn't appear in the cache until the next time something is posted.

A simple fix would be to have a cache expiration timer, but then the future posts would still be N minutes late to appear.

For now, removing this "optimization".