Integrating Next.js
Partially working implementation, still using a custom server (not Next API routes). URLs for pages are based on the same case as the file names, so links to localhost:3000/news don't work and need to be /News, for example.
Issue right now is getting a chain of dispatch() calls to work (thunk?) e.g. to fetch news.
Also note that this may not match some recent changes from the master branch (even before this check-in).