Newsletter Signup and Twitter Feed
Using Pages to create ways to embed external HTML; i.e. a mailchimp sign-up form on the News page, and a Twitter feed on the Home page.
You need to create the content for these in the Pages section of the site; one called Newsletter and another called TwitterFeed.

Removed some unused packages (in Communities, Fans, Vanarama), updated some others.